requirements.txt
flask==2.3.2
google-generativeai==0.8.6
boto3>=1.26.0
requests>=2.28.0
python-dotenv>=0.21.0
gunicorn>=20.1.0
aiohttp>=3.8.0
uvicorn>=0.20.0
pydantic>=2.0.0